requirements.txt
chardet
tqdm
lupa
pillow